requirements.txt
###rasa_nlu
klein==17.10.0
hyperlink==17.3.1
six==1.11.0
matplotlib==2.1.0
simplejson==3.13.2
cloudpickle==0.6.1


### Sklearn
spacy==2.0.18
sklearn-crfsuite==0.3.6

###rasa_core
## Setup py requirements
jsonpickle==1.0.0
redis==2.10.6
fakeredis==0.10.3
numpy==1.16.0
scipy==1.2.0
typing==3.6.6
ruamel.yaml==0.15.85
requests==2.21.0
apscheduler==3.5.3
tqdm==4.29.1
networkx==2.2
pykwalify==1.7.0
coloredlogs==10.0
flask==1.0.2
flask_cors==3.0.7
scikit-learn==0.20.2
rasa_nlu~=0.14.0
slackclient==1.3.0
python-telegram-bot==11.1.0
twilio==6.23.0
webexteamssdk==1.1.1
mattermostwrapper==2.1
colorhash==1.0.2
pika==0.12.0
jsonschema==2.6.0
packaging==18.0
gevent==1.4.0
pytz==2018.9
rasa_core_sdk~=0.12.1
pymongo==3.7.2
python-dateutil==2.7.5
rocketchat_API==0.6.25
colorclass==2.2.0
terminaltables==3.1.0
questionary==1.0.1
flask-jwt-simple==0.0.3
python-socketio==3.1.1
pydot==1.4.1
keras-applications==1.0.6
keras-preprocessing==1.0.5
protobuf==3.6.0
###additional
pypandoc
git+https://github.com/apixu/apixu-python.git
